+++ src/app/skus/list-sku/list-sku.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import {Router} from "@angular/router";
+import {DiscountService} from '../../service/discount.service';
 import {SkuService} from "../../service/sku.service";
 import {Sku} from "../../model/sku.model";
 
@@ -12,16 +13,19 @@ export class ListSkuComponent implements OnInit {
 
   skus: Sku[];
 
-  constructor(private router: Router, private skuService: SkuService) { }
+  constructor(private router: Router, private skuService: SkuService, private discountService: DiscountService) { }
 
   ngOnInit() {
     var data = this.skuService.getSkus()
     this.skus = data;
-  }
+  };
 
   deleteSku(sku: Sku): void {
-    this.skuService.deleteSku(sku.id)
-    this.skus = this.skus.filter(u => u !== sku);
+    if(window.confirm('Are you sure you want to delete this sku and all associated discounts?')){
+      this.discountService.deleteAllDiscountsInSku(sku.skuNumber);
+      this.skuService.deleteSku(sku.id)
+      this.skus = this.skus.filter(u => u !== sku);
+    }
   };
 
   editSku(sku: Sku): void {
@@ -38,7 +42,7 @@ export class ListSkuComponent implements OnInit {
     localStorage.removeItem("listSkuNumberForDiscounts");
     localStorage.setItem("listSkuNumberForDiscounts", sku.skuNumber);
     this.router.navigate(['list-discount']);
-  }
+  };
 
   menuClick(menuItem: string) {
     menuItem = menuItem.toUpperCase();
